CreateQuizApp splitted into more manegable chucnks
