+++ scripts/createQuizApp/data/sampleData.js
@@ -1,0 +1,712 @@
+module.exports = {
+    "Verbs3(Year3)": {
+        "name": "Verbs 3 (Year 3)",
+        "questions": [
+            {
+                "question": "I ______ dreams can come true.",
+                "answer": "believe",
+                "alternatives": [
+                    "behave",
+                    "tell",
+                    "talk"
+                ]
+            },
+            {
+                "question": "Builders _______ a new house next door.",
+                "answer": "built"
+            },
+            {
+                "question": "You may _______ to light a fire for warmth.",
+                "answer": "decide"
+            },
+            {
+                "question": "The class enjoys ________ new words.",
+                "answer": "learning"
+            },
+            {
+                "question": "We _______ what the thief looked like.",
+                "answer": "described"
+            },
+            {
+                "question": "Zofia was ________ to win the quiz!",
+                "answer": "surprised"
+            },
+            {
+                "question": "Glue makes glitter _______ to the paper.",
+                "answer": "stick"
+            },
+            {
+                "question": "The cat _______ the dog was a danger.",
+                "answer": "thought"
+            },
+            {
+                "question": "The way snowflakes ______ is interesting",
+                "answer": "form"
+            },
+            {
+                "question": "I _______ to always try my best at school!",
+                "answer": "promise"
+            }
+        ]
+    },
+    "TimesTable(1to6)": {
+        "name": "Times Table (1 to 6)",
+        "questions": [
+            {
+                "question": "5 x 2",
+                "answer": "10",
+                "alternatives": [
+                    "5",
+                    "12",
+                    "8"
+                ]
+            },
+            {
+                "question": "3 x 3",
+                "answer": "9"
+            },
+            {
+                "question": "4 x 6",
+                "answer": "24"
+            },
+            {
+                "question": "1 x 12",
+                "answer": "12"
+            },
+            {
+                "question": "6 x 7",
+                "answer": "42"
+            },
+            {
+                "question": "5 x 11",
+                "answer": "55"
+            },
+            {
+                "question": "2 x 8",
+                "answer": "16"
+            },
+            {
+                "question": "6 x 9",
+                "answer": "54"
+            },
+            {
+                "question": "3 x 12",
+                "answer": "36"
+            },
+            {
+                "question": "6 x 5",
+                "answer": "30"
+            }
+        ]
+    },
+    "ArtistsandPaintings": {
+        "name": "Artists and Paintings",
+        "questions": [
+            {
+                "question": "Who painted the Mona Lisa?",
+                "answer": "Leonardo Da Vinci"
+            },
+            {
+                "question": "Who painted The Scream?",
+                "answer": "Edward Munch"
+            },
+            {
+                "question": "Who painted The Starry Night ",
+                "answer": "Vincent van Gogh"
+            },
+            {
+                "question": "Who painted Guernica?",
+                "answer": "Pablo Picasso"
+            },
+            {
+                "question": "Who painted Girl with a Pearl Earring?",
+                "answer": "Johannes Vermeer"
+            },
+            {
+                "question": "Who painted The Night Watch?",
+                "answer": "Rembrandt"
+            },
+            {
+                "question": "Who painted The Creation of Adam in the Sistine Chapel?",
+                "answer": "Mihcelangelo"
+            },
+            {
+                "question": "Who painted The Persistence of Memory",
+                "answer": "Salvador Dali"
+            },
+            {
+                "question": "Who painted Water Lilies?",
+                "answer": "Claude Monet"
+            },
+            {
+                "question": "Who painted Bathers at Asnières?",
+                "answer": "George Seurat"
+            }
+        ]
+    },
+    "Shapes(2D)": {
+        "name": "Shapes (2D)",
+        "questions": [
+            {
+                "question": "What shape has four equal sides?",
+                "answer": "Square"
+            },
+            {
+                "question": "What shape has six sides?",
+                "answer": "Hexagon"
+            },
+            {
+                "question": "What shape has three equal sides",
+                "answer": "Equilateral Triangle"
+            },
+            {
+                "question": "What shape has eight sides?",
+                "answer": "Octagon"
+            },
+            {
+                "question": "What shape has ten sides?",
+                "answer": "Decagon"
+            },
+            {
+                "question": "What shape has four equal angles but not four equal sides? ",
+                "answer": "Rectangle"
+            },
+            {
+                "question": "What shape has five sides?",
+                "answer": "Pentagon"
+            },
+            {
+                "question": "What shape three unequal sides?",
+                "answer": "Scalene Triangle"
+            },
+            {
+                "question": "What shape has nine sides?",
+                "answer": "Nonagon"
+            },
+            {
+                "question": "What shape has seven sides?",
+                "answer": "Septagon"
+            }
+        ]
+    },
+    "EuropeanGeography(CapitalCities)": {
+        "name": "European Geography (Capital Cities)",
+        "questions": [
+            {
+                "question": "What is the captial of Germany?",
+                "answer": "Berlin"
+            },
+            {
+                "question": "What is the capital of Greece?",
+                "answer": "Athens"
+            },
+            {
+                "question": "What is the capital of Portugal?",
+                "answer": "Lisbon"
+            },
+            {
+                "question": "What is the capital of Russia?",
+                "answer": "Moscow"
+            },
+            {
+                "question": "What is the capital of Sweden?",
+                "answer": "Stockholm"
+            },
+            {
+                "question": "What is the captial of Ukraine?",
+                "answer": "Kiev"
+            },
+            {
+                "question": "What is the capital of Austria?",
+                "answer": "Vienna"
+            },
+            {
+                "question": "What is the capital of Poland?",
+                "answer": "Warsaw"
+            },
+            {
+                "question": "What is the captial of Norway?",
+                "answer": "Oslo"
+            },
+            {
+                "question": "What is the capital of Switzerland?",
+                "answer": "Bern"
+            }
+        ]
+    },
+    "BritishHistory(WWI)": {
+        "name": "British History (WWI)",
+        "questions": [
+            {
+                "question": "The assassination of whom triggered WWI?",
+                "answer": "Franz Ferdinand",
+                "alternatives": [
+                    "George V",
+                    "Nicholas II",
+                    "Wilhelm II"
+                ]
+            },
+            {
+                "question": "Who was the King of the United Kingdom?",
+                "answer": "George V"
+            },
+            {
+                "question": "What was the name of the coalition fighting the Allied Powers?",
+                "answer": "Central Powers"
+            },
+            {
+                "question": "Who was the Emperor of Russia? ",
+                "answer": "Nicholas II"
+            },
+            {
+                "question": "Who was the British Prime Minister at the end of the war?",
+                "answer": "David Lloyd George"
+            },
+            {
+                "question": "Which Empire had the biggest army in the war?",
+                "answer": "German Empire"
+            },
+            {
+                "question": "Which Asian country fought with the Allied Powers?",
+                "answer": "Japan"
+            },
+            {
+                "question": "Who was the Emperor of Germany?",
+                "answer": "Wilhelm II"
+            },
+            {
+                "question": "Which member of the Allied Powers suffered the most casualties?",
+                "answer": "Russia"
+            },
+            {
+                "question": "Who was the American President during WWI?",
+                "answer": "Woodrow Wilson"
+            }
+        ]
+    },
+    "Spanish(Animals)": {
+        "name": "Spanish (Animals)",
+        "questions": [
+            {
+                "question": "What is the Spanish for 'dog'?",
+                "answer": "Perro"
+            },
+            {
+                "question": "What is the Spanish for 'cat'?",
+                "answer": "Gato"
+            },
+            {
+                "question": "What is the Spanish for 'butterfly'?",
+                "answer": "Mariposa"
+            },
+            {
+                "question": "What is the Spanish for 'cow'?",
+                "answer": "Vaca"
+            },
+            {
+                "question": "What is the Spanish for 'monkey'?",
+                "answer": "Mono"
+            },
+            {
+                "question": "What is the Spanish for 'chicken'?",
+                "answer": "Pollo"
+            },
+            {
+                "question": "What is the Spanish for 'swan'?",
+                "answer": "Cisne"
+            },
+            {
+                "question": "What is the Spanish for 'bird'?",
+                "answer": "Pájaro"
+            },
+            {
+                "question": "What is the Spanish for 'horse'?",
+                "answer": "Caballo"
+            },
+            {
+                "question": "What is the Spanish for 'bee'?",
+                "answer": "Abeja"
+            }
+        ]
+    },
+    "French(Animals)": {
+        "name": "French (Animals)",
+        "questions": [
+            {
+                "question": "What is the French for 'butterly'?",
+                "answer": "Papillon"
+            },
+            {
+                "question": "What is the French for 'cow'?",
+                "answer": "Vache"
+            },
+            {
+                "question": "What is the French for 'dog'?",
+                "answer": "Chien"
+            },
+            {
+                "question": "What is the French for 'chicken'?",
+                "answer": "Poulet"
+            },
+            {
+                "question": "What is the French for 'cat'? ",
+                "answer": "Chat"
+            },
+            {
+                "question": "What is the French for 'horse'?",
+                "answer": "Cheval"
+            },
+            {
+                "question": "What is the French for 'swan'?",
+                "answer": "Cygne"
+            },
+            {
+                "question": "What is the French for 'monkey'?",
+                "answer": "Singe"
+            },
+            {
+                "question": "What is the French for 'bee'?",
+                "answer": "Abeille"
+            },
+            {
+                "question": "What is the French for 'giraffe'?",
+                "answer": "Girafe"
+            }
+        ]
+    },
+    "Verbs5(Year5)": {
+        "name": "Verbs 5 (Year 5)",
+        "questions": [
+            {
+                "question": "We were well _______ for the adventure.",
+                "answer": "equipped",
+                "alternatives": [
+                    "quitted",
+                    "equated",
+                    "quoted"
+                ]
+            },
+            {
+                "question": "The netball captain was _________ by the loss.",
+                "answer": "embarrassed"
+            },
+            {
+                "question": "I want to ________ our cat to the vet.",
+                "answer": "accompany"
+            },
+            {
+                "question": "The police _______ to know the proper law.",
+                "answer": "ought"
+            },
+            {
+                "question": "Michael always ________ the teacher in class!",
+                "answer": "interrupts"
+            },
+            {
+                "question": "My sister ________ an A grade in her maths exam.",
+                "answer": "achieved"
+            },
+            {
+                "question": "The doctor ______ a severe case of measles.",
+                "answer": "diagnosed"
+            },
+            {
+                "question": "I really _________ having friends to help me.",
+                "answer": "appreciate"
+            },
+            {
+                "question": "Our teacher _______ us to attempt hard questions.",
+                "answer": "encourages"
+            },
+            {
+                "question": "Birds of prey _______ down on smaller animals.",
+                "answer": "swoop"
+            }
+        ]
+    },
+    "USGeography(StateCapitals)": {
+        "name": "US Geography (State Capitals)",
+        "questions": [
+            {
+                "question": "What is the capital of Arizona?",
+                "answer": "Phoenix"
+            },
+            {
+                "question": "What is the capital of Colorado?",
+                "answer": "Denver"
+            },
+            {
+                "question": "What is the capital of Georgia?",
+                "answer": "Atlanta"
+            },
+            {
+                "question": "What is the capital of Hawaii?",
+                "answer": "Honolulu"
+            },
+            {
+                "question": "What is the capital of Indiana?",
+                "answer": "Indianapolis"
+            },
+            {
+                "question": "What is the capital of Massachusetts?",
+                "answer": "Boston"
+            },
+            {
+                "question": "What is the capital of New Mexico?",
+                "answer": "Santa Fe"
+            },
+            {
+                "question": "What is the capital of Ohio?",
+                "answer": "Columbus"
+            },
+            {
+                "question": "What is the capital of Tennessee?",
+                "answer": "Nashville"
+            },
+            {
+                "question": "What is the capital of Texas?",
+                "answer": "Austin"
+            }
+        ]
+    },
+    "InventorsandInventions": {
+        "name": "Inventors and Inventions",
+        "questions": [
+            {
+                "question": "Who invented the electric light bulb?",
+                "answer": "Thomas Edison"
+            },
+            {
+                "question": "Who invented the first powered areoplane?",
+                "answer": "The Wright Brothers"
+            },
+            {
+                "question": "Who invented the first working telephone?",
+                "answer": "Alexander Graham Bell"
+            },
+            {
+                "question": "Who invented the world wide web?",
+                "answer": "Tim Berners Lee"
+            },
+            {
+                "question": "Who invented the first vaccine?",
+                "answer": "Edward Jenner"
+            },
+            {
+                "question": "Who invented the television?",
+                "answer": "John Logie Baird"
+            },
+            {
+                "question": "Who invented the first real computer?",
+                "answer": "Alan Turing"
+            },
+            {
+                "question": "Who invented the steam engine?",
+                "answer": "James Watt"
+            },
+            {
+                "question": "Who invented the hydrogen bomb?",
+                "answer": "Edward Teller"
+            },
+            {
+                "question": "Who invented the first real battery?",
+                "answer": "Alessandro Volta"
+            }
+        ]
+    },
+    "SimpleAlgebra": {
+        "name": "Simple Algebra",
+        "questions": [
+            {
+                "question": "n6 = 18",
+                "answer": "n = 3"
+            },
+            {
+                "question": "n2 + 3 = 11",
+                "answer": "n = 4"
+            },
+            {
+                "question": "7 - 1n = 0",
+                "answer": "n = 7"
+            },
+            {
+                "question": "10 + 5n = 35",
+                "answer": "n = 5"
+            },
+            {
+                "question": "2n x 3n = 5",
+                "answer": "n = 1"
+            },
+            {
+                "question": "n4 + 10 = 34",
+                "answer": "n = 6"
+            },
+            {
+                "question": "5 + n3 = 5",
+                "answer": "n = 0"
+            },
+            {
+                "question": "8n + 20 = 100",
+                "answer": "n = 10"
+            },
+            {
+                "question": "5n + 5 = 45",
+                "answer": "n = 8"
+            },
+            {
+                "question": "8 + 3n = 14",
+                "answer": "n = 2"
+            }
+        ]
+    },
+    "Biology(HumanOrgans)": {
+        "name": "Biology (Human Organs)",
+        "questions": [
+            {
+                "question": "An organ that pumps blood around the body?",
+                "answer": "Heart",
+                "alternatives": [
+                    "Kidney",
+                    "Lung",
+                    "Bladder"
+                ]
+            },
+            {
+                "question": "A sense organ used in touch?",
+                "answer": "Skin"
+            },
+            {
+                "question": "An organ that fitlers blood?",
+                "answer": "Kidney"
+            },
+            {
+                "question": "An organ that produces insulin?",
+                "answer": "Pancreas"
+            },
+            {
+                "question": "A sense organ used in sight?",
+                "answer": "Eye"
+            },
+            {
+                "question": "An organ that produces bile?",
+                "answer": "Liver"
+            },
+            {
+                "question": "An organ used in respiration?",
+                "answer": "Lung"
+            },
+            {
+                "question": "A female reproductive organ?",
+                "answer": "Ovary"
+            },
+            {
+                "question": "A sense organ used in smell?",
+                "answer": "Nose"
+            },
+            {
+                "question": "An organ that collects urine?",
+                "answer": "Bladder"
+            }
+        ]
+    },
+    "BritishHistory(TheTudors)": {
+        "name": "British History (The Tudors)",
+        "questions": [
+            {
+                "question": "Who defeated Richard III to end the War of the Roses",
+                "answer": "Henry VII ",
+                "alternatives": [
+                    "Henry VIII",
+                    "Henry VI",
+                    "Henry V"
+                ]
+            },
+            {
+                "question": "Which son of Henry Tudor had six wives?",
+                "answer": "Henry VIII"
+            },
+            {
+                "question": "Who was the first wife of Henry VIII?",
+                "answer": "Catherine of Aragon"
+            },
+            {
+                "question": "Which queen was the first daughter of Henry VIII?",
+                "answer": "Mary I"
+            },
+            {
+                "question": "Who took the throne after Henry VIII died?",
+                "answer": "Edward VI"
+            },
+            {
+                "question": "Which future queen was the sister of Mary I?",
+                "answer": "Elizabeth I"
+            },
+            {
+                "question": "Who was Elizabeth I's mother?",
+                "answer": "Anne Boleyn"
+            },
+            {
+                "question": "Which Catholic king did Mary I marry?",
+                "answer": "Philip of Spain"
+            },
+            {
+                "question": "Which queen did Elizabeth I have imprisoned?",
+                "answer": "Mary Queen of Scots"
+            },
+            {
+                "question": "Which Scottish king James I of England?",
+                "answer": "James VI of Scotland"
+            }
+        ]
+    },
+    "TimesTable(7to12)": {
+        "name": "Times Table (7 to 12)",
+        "questions": [
+            {
+                "question": "7 x 6",
+                "answer": "42",
+                "alternatives": [
+                    "46",
+                    "38",
+                    "32"
+                ]
+            },
+            {
+                "question": "12 x 3",
+                "answer": "36"
+            },
+            {
+                "question": "8 x 9",
+                "answer": "72"
+            },
+            {
+                "question": "11 x 5",
+                "answer": "55"
+            },
+            {
+                "question": "8 x 4",
+                "answer": "32"
+            },
+            {
+                "question": "9 x 6",
+                "answer": "54"
+            },
+            {
+                "question": "10 x 11",
+                "answer": "110"
+            },
+            {
+                "question": "12 x 12",
+                "answer": "144"
+            },
+            {
+                "question": "9 x 10",
+                "answer": "90"
+            },
+            {
+                "question": "7 x 4",
+                "answer": "28"
+            }
+        ]
+    }
+};
