+++ scripts/quiz.js
@@ -7,7 +7,7 @@ function login() {
   token = localStorage.getItem("zzishtoken");
   email = localStorage.getItem("emailAddress");
   if (token==null && email==null) {
-      Zzish.login(type,url,function(err,message) {          
+      Zzish.login(type,url,function(err,message) {
           $("#LoginButton").html("Logout");
           console.log("Logged in with status and message",err,message);
           if (!err) {
@@ -16,7 +16,7 @@ function login() {
           else {
               console.log("Error",err);
           }
-      });            
+      });
    }
    else if (email!=null) {
           localStorage.clear();
@@ -24,17 +24,17 @@ function login() {
           location.href="/quiz/";
    }
    else {
-      Zzish.logout(token,function(err,message) {        
+      Zzish.logout(token,function(err,message) {
           localStorage.clear();
           $("#LoginButton").html("Login with Zzish");
           location.href="/quiz/";
       });
-  }                
-} 
+  }
+}
 
 $( document ).ready(function() {
-	quizData = localStorage.getItem("quizData");
-	emailAddress = localStorage.getItem("emailAddress");
+	var quizData = localStorage.getItem("quizData");
+	var emailAddress = localStorage.getItem("emailAddress");
 	if (quizData!=undefined) {
 		qj = $.parseJSON(quizData);
 		if (qj!=undefined && qj.length>0) {
