+++ scripts/quizApp/config/settings.js
