+++ scripts/quizApp/controllers/IntroController.js
