+++ scripts/quizApp/controllers/ManualController.js
