+++ scripts/quizApp/controllers/ScrambledController.js
