+++ scripts/quizApp/controllers/StartController.js
