+++ scripts/quizApp/utils/settings.js
