+++ views/create.jade
@@ -9,23 +9,21 @@ html(ng-app='createQuizApp')
         <meta name="apple-mobile-web-app-capable" content="yes" />
         <meta name="apple-mobile-web-app-status-bar-style" content="black">
 
-        script(src='/js/jquery-1.11.0.js')
-        script(src='/js/fastclick.js')
-        script(src='/js/uuid.js')
-        script(src='/js/bootstrap.js')
+        script(src='/lib/jquery-1.11.0.js')
+        script(src='/lib/fastclick.js')
+        script(src='/lib/uuid.js')
+        script(src='/lib/bootstrap.js')
+        script(src='/LoginButton/jquery-ui.min.js')  
 
         <!--[if lt IE 9]>
         script(src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js")
         script(src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js")
         <![endif]-->
 
-        script(src='/js/angular.min.js')
-        script(src='/js/angular-route.min.js')
-        script(src='/js/angular-animate.min.js')
         script(src='#{session.zzishsdkurl}/zzish.js')        
+        script(src='/js/vendor.js')        
         script(src='/js/createQuizApp.js')        
         script(src='/js/quiz.js')  
-        script(src='/js/jquery-ui.min.js')  
 
 
 
@@ -202,4 +200,3 @@ body(ng-controller="NavBarController as ctrl")
                         br
                         br
                         button.btn.btn-info.btn-block(ng-click="ctrl.dismiss();") Get Started
-
